+++ pkg/grafana/types.go
@@ -44,12 +44,9 @@ func (d Datasource) buildJSONData() map[string]interface{} {
 
 func (d Datasource) buildSecureJSONData(organization Organization) map[string]string {
 	tenantIDs := organization.TenantIDs
-	if d.UID == mimirDatasourceUID {
-		// We do not support multi-tenancy for Mimir yet
-		tenantIDs = common.DefaultReadTenants
-	} else if d.UID == mimirAlertmanagerDatasourceUID {
-		// Alertmanager does not support multi-tenancy
-		tenantIDs = []string{common.DefaultWriteTenant}
+	if d.UID == mimirOldDatasourceUID {
+		// For the old Mimir datasource, we need to use the anonymous tenant
+		tenantIDs = []string{"anonymous"}
 	}
 	return map[string]string{
 		"httpHeaderValue1": strings.Join(tenantIDs, "|"),